+++ src/FSharp/Random.fs
@@ -39,11 +39,11 @@ module Random =
     /// Default sampling, efficient but without custom seed (uses robust seeds internally)
     let inline doubles (length:int) = SystemRandomSource.FastDoubles(length)
     let inline doubleSeq () = SystemRandomSource.DoubleSequence()
-    let inline doubleFill (values:double[]) = SystemRandomSource.FastDoubles(values)
+    let inline doubleFill (values:float[]) = SystemRandomSource.FastDoubles(values)
 
     let inline doublesSeed (seed:int) (length:int) = SystemRandomSource.Doubles(length, seed)
     let inline doubleSeqSeed (seed:int) = SystemRandomSource.DoubleSequence(seed)
-    let inline doubleFillSeed (seed:int) (values:double[]) = SystemRandomSource.Doubles(values, seed)
+    let inline doubleFillSeed (seed:int) (values:float[]) = SystemRandomSource.Doubles(values, seed)
 
     /// Creates a default .Net system pRNG with a robust seed
     let systemShared = shared
